refactor(timer): abstract logic into separate class
The app can now be made fully multiplatform once ServiceHelper is migrated
