+++ androidApp/src/main/java/org/nsh07/pomodoro/di/androidModules.kt
@@ -22,6 +22,7 @@ import android.app.PendingIntent
 import android.content.Context
 import android.content.Context.NOTIFICATION_SERVICE
 import android.os.Build
+import android.os.SystemClock
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
 import androidx.core.app.NotificationCompat
@@ -41,6 +42,7 @@ import org.nsh07.pomodoro.data.StatRepository
 import org.nsh07.pomodoro.data.StateRepository
 import org.nsh07.pomodoro.service.AppServiceHelper
 import org.nsh07.pomodoro.service.ServiceHelper
+import org.nsh07.pomodoro.service.TimerManager
 import org.nsh07.pomodoro.service.addTimerActions
 
 val servicesModule = module {
@@ -51,6 +53,7 @@ val servicesModule = module {
     single<AppPreferenceRepository>() bind PreferenceRepository::class
     single<StateRepository>()
     single<AppServiceHelper>() bind ServiceHelper::class
+    single<TimerManager> { TimerManager(get(), get(), SystemClock::elapsedRealtime) }
 
     single { NotificationManagerCompat.from(get()) }
     single { create(::createNotificationManager) }
@@ -89,7 +92,7 @@ private fun createNotificationCompatBuilder(context: Context): NotificationCompa
                 PendingIntent.FLAG_IMMUTABLE
             )
         )
-        .addTimerActions(context, R.drawable.play, context.getString(R.string.start))
+        .addTimerActions(context, context.getString(R.string.start))
         .setShowWhen(true)
         .setSilent(true)
         .setOngoing(true)
